backend/requirements.txt
fastapi
httpx
python-dotenv
uvicorn
websockets
